treewide: embed va_list directly in struct va_format
This saves 8 bytes of stack and a bit of code in the affected
functions.

But probably not worth the churn, total .text saving for defconfig
x86_64 is a mere 800 bytes.

Mostly generated by this semantic patch:

@r depends on patch exists@
identifier ap, vaf, last;
@@
  struct va_format vaf;
  ...
- va_list ap;
  ...
- va_start(ap, last);
  ...
- vaf.va = &ap;
+ va_start(vaf.va, last);
  ...
- va_end(ap);
+ va_end(vaf.va);

@r2 depends on patch exists@
identifier ap, vaf, last;
@@
- va_list ap;
  ...
  struct va_format vaf;
  ...
- va_start(ap, last);
  ...
- vaf.va = &ap;
+ va_start(vaf.va, last);
  ...
- va_end(ap);
+ va_end(vaf.va);

This needs to be run with a --macro-file argument pointing to a file
which has at least

#define __printf(a, b)
#define __cold

to make spatch actually parse functions with those attributes attached
at their definition site.

Obviously, on top of that, fixing the definition of struct va_format and
its use in vsprintf.c, a few places that does *vaf->va for passing to
vsnprintf, and some places that the .cocci script didn't find.