RFC/hack: out-of-line va_arg functions
On x86-64, and presumably other architectures with a complicated scheme
for passing varargs (basically, anything that passes some in registers),
va_arg() expands inline to quite a bit of code. On those same
architectures, va_list is typically a typedef for a one-element array of
some "control structure", which is effectively passed by reference. That
allows us to define out-of-line versions of the various va_arg calls,
and choose the appropriate one with some nested __builtin_choose_expr
and __builtin_types_compatible_p.

The implementation assumes that the calling convention for passing a
pointer is the same as for passing a long - moreover, we assume that a
long is passed the same way as a long long (64 bit) or int (32 bit). The
latter is obviously easy to relax. The former could probably also be
relaxed (as long as all pointers are passed using the same convention as
void*).

This explicitly does not work for compound types, but I highly doubt we
pass a struct by value to a variadic function. But just in case, using
these should be opt-in per translation unit, so I've added a guard to
prevent linux/stdarg.h from being included by some other header.

This is of course mainly of use in lib/vsprintf.c, which by far has the
largest number of va_arg invocations.