jump_label: move HAVE_JUMP_LABEL to new asm-generic file
I'm going to need to know whether HAVE_JUMP_LABEL is defined in some
assembly code (this is for x86, and while we require CC_HAVE_ASM_GOTO,
CONFIG_JUMP_LABEL may not be set). So move the definition of
HAVE_JUMP_LABEL to an asm-generic header that can safely be included
from arch asm headers.

Similarly, I'm going to need to refer to the JUMP_TYPE_FALSE and
JUMP_TYPE_TRUE constants from asm code. So also move those and define
them using the UL() helper macro, but only define them in the
HAVE_JUMP_LABEL case, just as they used to be.

Cc: Jason Baron <jbaron@akamai.com>