parse_integer: sane integer parsing
Based on original idea by Alexey Dobriyan, this implements a
type-generic parse_integer function (well, actually macro) with
reasonably sane semantics. On success, the number of consumed
characters is consumed, and on error, a negative Exxx value is
returned (-EINVAL or -ERANGE).

This is supposed to be "one interface to rule them all", so one can
modify the behaviour with a number of flags. Say, if one wants the
skip-leading-whitespace behaviour of strtol() and friends, just set
PARSE_INTEGER_LEAD_WS. If you want the entire string to represent an
integer and nothing else (except perhaps for the whitespace you allow
with other flags), use PARSE_INTEGER_ALL to make it do the check for
'\0' for you. If an insanely large value should just be treated as
whatever maximum the type holds, you can use PARSE_INTEGER_SATURATE to
alter the behaviour in case of ERANGE. Or PARSE_INTEGER_WRAP to get
modulo 2^N behaviour.

Of course, I don't intend for each driver to choose their own random
collection of PARSE_INTEGER_* flags, but the idea is that (1) we can
get rid of a number of the existing interfaces (kstrtox) - that alone
is a net .text saving, and (2) when somebody somewhere needed
semantics not _quite_ covered by one of the existing families, they
may have added their own private integer parsing routine - hopefully
such places can also be replaced with an appropriate
parse_integer(...).

The implementation assumes a sane environment (2's complement, no
padding bits, no nonsense). I only distinguish integers based on their
width and signedness, so yes, this means that I may end up, for
example, accessing a long object as if it was an int or vice versa.